requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
beautifulsoup4==4.12.2
pydantic==2.5.2
python-multipart==0.0.6 